refactor(backend): extract subscription service from tier middleware
